requirements.txt
tensorflow >= 1.12
tfserve
numpy
pillow
